pyproject.toml
[project]
name = "crm-ai-agent"
version = "1.0.0"
description = "LangGraph-based parsing service for extracting structured CRM data from unstructured interactions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Core dependencies
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    # LangGraph and LangChain
    "langgraph==0.0.20",
    "langchain>=0.1.0",
    "langchain-google-vertexai==0.0.6",
    "langchain-core>=0.1.23,<0.2",
    # Google Cloud
    "google-cloud-bigquery>=3.13.0",
    "google-cloud-pubsub>=2.18.4",
    "google-cloud-storage>=2.14.0",
    "google-cloud-speech>=2.21.0",
    "google-cloud-aiplatform>=1.42.1",
    # Audio processing
    "openai-whisper==20231117",
    # Utilities
    "python-dateutil==2.8.2",
    "python-multipart==0.0.6",
    "httpx==0.25.2",
    "langdetect==1.0.9",
    "langchain-google-community>=1.0.4",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "models", "services"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

